pyproject.toml
# Project metadata and tool configurationi
# Configures project metadata and tools like Black, Flake8, and isort.
[project]
name = "enterprise_app"
version = "1.0.0"
description = "Enterprise-grade Python web application with automation capabilities."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.30.6",
    "pydantic==2.9.2",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "loguru==0.7.2",
    "pandas==2.2.3",
    "sqlalchemy==2.0.35",
    "asyncpg==0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "black==24.8.0",
    "flake8==7.1.1",
    "isort==5.13.2",
    "pre-commit==3.8.0",
    "coverage==7.6.1",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
